Separate generic-rfq e2e tests from local e2e tests.
